// web/types/fomo.ts
export interface FomoDependencyItem {
  id?: string;
  projectId?: string;
  project_id?: string;
  slug?: string;
  title?: string;
  name?: string;
  description?: string;
  author?: string;
  categories?: any[];
  iconUrl?: string | null;
  icon_url?: string | null;
  dependencyType?: string;
  dependency_type?: string;
  projectType?: string;
  project_type?: string;
  versionId?: string | null;
  version_id?: string | null;
  fileName?: string | null;
  file_name?: string | null;
  url?: string | null;
  _source?: string;
  [key: string]: any;
}